Code refactoring
* Renamed TimeRef to TTimeRef, in future all Classes will have a
preceding T to improve readability
* Moved TTimeRef to it's own file instead of defines.h making it more
clear which class actually need TTimeRef

* Improve readability on which function is called (re-implemented or
base virtual function). No functionality change here

* Using QString multi arg, it's more efficient

* First start to move common code to LocationItem for classes that have
a place on the timeline, like AudioClip, Curve, Marker. In the future
SnapList will use LocationItem for snapping so move to snap can be done
globally or per class using same snap logic

* Tsar
- Sloppy coding style fixed, members now have a preceding m_ as well as
renaming couple of variables to make it more clear what their purpose
is.

- Adding events to the event buffer logic assumed the buffer write
always worked. The buffer is so large and is being emptied so often this
should be true, but no harm done in actually checking this and notifying
user that the event buffer ran out of space. IN that case the user did
something unexpected or the program needs to be fixed for that
particular use case.

- Moved thread_save_* defines into functions, according to my colleague
Sadok (and he knows his stuff) it's just silly to define a function that
the pre-processor puts everywhere in the code base then just define it
once in Tsar as a proper function.

* struct transort_state_t -> class TTransortControl
and put it into it's own dedicated file including proper setter/getter
functions for all it's members
The AudioDevice and JackDriver own a TTransportControl object now and
are passed to functions by pointer so it's not getting out of scope and
being implicitely copied by passing them to functions.

Reason: Before it was not possible to track changes in transort control
states, now we can actually start tracking changes in transport control
changes since only 1 instance exisist. The logic for tracking changes
still needs to be added though.

* ItemBrowserData
used for keyboard based cursor placemented, or navigating through the
scene by keyboard. It was quickly hacked and clang tidy reminded of the
clopy coding. For now added couple of ASSERTS to catch real problems. As
with placments of Playhead/Work cursor this logic needs to be reviewed
anyways

+++ src/traverso/widgets/SpectralMeterWidget.cpp
@@ -193,7 +193,7 @@ void SpectralMeterView::update_background()
 	QFontMetrics fm(themer()->get_font("FFTMeter:fontscale:label"));
 
 	// draw horizontal lines + labels
-	for (float i = upper_db; i >= lower_db; i -= 10.0f) {
+    for (int i = int(upper_db); i >= int(lower_db); i -= 10) {
 		float f = db2ypos(i);
 
 		painter.setPen(m_penGrid);
